+++ src/main/java/tornadofx/Workspace.kt
@@ -218,7 +218,7 @@ open class Workspace(title: String = "Workspace", navigationMode: NavigationMode
 
     val header: ToolBar get() = root.header
 
-    private fun navigateForward(): Boolean {
+    fun navigateForward(): Boolean {
         if (!forwardButton.isDisabled) {
             dock(viewStack[viewPos.get() + 1], false)
             return true