+++ src/test/kotlin/tornadofx/tests/StylesheetTest.kt
@@ -19,6 +19,7 @@ import tornadofx.Stylesheet.Companion.barChart
 import tornadofx.Stylesheet.Companion.hover
 import tornadofx.Stylesheet.Companion.imageView
 import tornadofx.Stylesheet.Companion.label
+import tornadofx.Stylesheet.Companion.pannable
 import tornadofx.Stylesheet.Companion.star
 import java.net.URI
 import kotlin.test.assertEquals
@@ -427,10 +428,10 @@ class StylesheetTest {
             }
         } shouldEqual {
             """
-        .label:hover, .label:armed, .text:hover, .text:armed {
-            -fx-background-color: rgba(0, 0, 255, 0.25);
-        }
-        """
+            .label:hover, .label:armed, .text:hover, .text:armed {
+                -fx-background-color: rgba(0, 0, 255, 0.25);
+            }
+            """
         }
     }
 
@@ -537,7 +538,7 @@ class StylesheetTest {
     @Test
     fun barFillTest() {
         stylesheet {
-            s(barChart) {
+            barChart {
                 barFill = Color.RED
             }
         } shouldEqual {
@@ -549,9 +550,30 @@ class StylesheetTest {
         }
     }
 
+    @Test
+    fun testNameCollisions() {
+        stylesheet {
+            pannable {
+                pannable = false
+                pannable {
+                    pannable = true
+                }
+            }
+        } shouldEqual {
+            """
+            :pannable {
+                -fx-pannable: false;
+            }
+            :pannable :pannable {
+                -fx-pannable: true;
+            }
+            """
+        }
+    }
+
     private fun stylesheet(op: Stylesheet.() -> Unit) = Stylesheet().apply(op)
-    infix fun Stylesheet.shouldEqual(op: () -> String) = Assert.assertEquals(op().strip(), render().strip())
-    infix fun Color.shouldEqual(other: Color) = Assert.assertEquals(other.toString(), toString())
+    private infix fun Stylesheet.shouldEqual(op: () -> String) = Assert.assertEquals(op().strip(), render().strip())
+    private infix fun Color.shouldEqual(other: Color) = Assert.assertEquals(other.toString(), toString())
     private fun String.strip() = replace(Regex("\\s+"), " ").trim()
 
     /**