+++ CHANGELOG.md
@@ -9,10 +9,12 @@
 
 ### Additions
 
+- `readonly` and `cancel` pseudoclasses added to type safe CSS
+
 - ContextMenu.radiomenuitem (https://github.com/edvin/tornadofx/issues/646)
 - mutableList can now be bound to an ObservableMap
 ## [1.7.15]
-
+y
 ### Fixed
 
 - onEditCommit {} listens for changes in nested columns