+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changes
 
+- Kotlin 1.2.30
 - Workspace.navigateForward() made public
 - Added missing pseudoclasses to CSS DSL